Re-organize provider tests into separate files
